[ISSUE #16] Set up Maven CI & Add .asf.yaml & Rename core packages (#17)
* Set up CI

* Add LICENSE

* Add CheckStyle

* Set up DB

* Add druid StatFilter controller

* sync CI change in https://github.com/apache/eventmesh/pull/4722

* Supplement module dependency relation and fix project info

* Move original Admin into Core & rewrite dependencies

* Rename package to .core

* Add .asf.yaml

* Enable branches with namespace triggering CI

* Set up Maven CI

* Downgrade JDK from 17 to 8

* Remove failed empty test

* Remove failed dependabot task

* Unify version number (using maven 3.8.1, maven wrapper to be added)

---------
