+++ .asf.yaml
@@ -1,0 +1,55 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+github:
+  description: EventMesh Dashboard
+  features:
+    # Enable issue management
+    issues: true
+    # Enable wiki
+    wiki: true
+  homepage: https://eventmesh.apache.org/
+  labels:
+    - pubsub
+    - event-mesh
+    - admin
+    - dashboard
+    - event-driven
+    - cloud-native
+    - serverless
+    - serverless-workflow
+    - message-bus
+  enabled_merge_buttons:
+    squash:  true
+    merge:   false
+    rebase:  false
+  protected_branches:
+    main:
+      required_status_checks:
+        strict: true
+      required_pull_request_reviews:
+        dismiss_stale_reviews: true
+        required_approving_review_count: 1
+notifications:
+    commits:      commits@eventmesh.apache.org
+    # Send all issue emails (new, closed, comments) to issues@
+    issues:       issues@eventmesh.apache.org
+    # Send new/closed PR notifications to dev@
+    pullrequests_status: dev@eventmesh.apache.org
+    # Send individual PR comments/reviews to issues@
+    pullrequests_comment: issues@eventmesh.apache.org
+    jira_options: link label worklog
