+++ .github/workflows/ci.yml
@@ -21,9 +21,9 @@ name: "Continuous Integration"
 
 on:
   push:
-    branches: [ '*' ]
+    branches: [ '**' ]
   pull_request:
-    branches: [ '*' ]
+    branches: [ '**' ]
 
 jobs:
   build:
@@ -32,7 +32,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ ubuntu-latest ]
-        java: [ 8 ]
+        java: [ 8, 11 ]
         language: ['java']
     runs-on: ${{ matrix.os }}
 
@@ -40,16 +40,12 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v3
 
-      - name: Setup Gradle
-        uses: gradle/gradle-build-action@v2
-
       - name: Set up JDK ${{ matrix.java }}
         uses: actions/setup-java@v3
         with:
           distribution: 'corretto'
           java-version: ${{ matrix.java }}
+          cache: maven
 
-      - name: Build
-        run: ./gradlew clean build bootJar --parallel --daemon
-        env:
-          GRADLE_ENTERPRISE_ACCESS_KEY: ${{ secrets.GE_ACCESS_TOKEN }}
+      - name: Build with Maven
+        run: mvn -B package --file pom.xml
