+++ README.md
@@ -6,6 +6,12 @@
 
 ## 技术架构
 
+### 环境
+
+- JDK 8/11
+- Maven 3.8.1
+- Spring Boot 2.7.x
+
 ### 模块依赖图
 
 ### 模块介绍
