+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/model/ConnectionInfo.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.model;
+package org.apache.eventmesh.dashboard.core.model;
 
 public class ConnectionInfo {
