+++ pom.xml
@@ -34,8 +34,7 @@
 
   <groupId>org.apache.eventmesh.dashboard</groupId>
   <artifactId>eventmesh-dashboard</artifactId>
-
-  <version>0.0.1</version>
+  <version>0.0.1-SNAPSHOT</version>
 
   <name>eventmesh-dashboard</name>
   <description>eventmesh-dashboard</description>
